Fix warnings (The uninitialised loop variable couald be really bad!)

+++ weaponObjectMachine/StateMachine.h
@@ -31,8 +31,8 @@ class StateMachineClass
     void setMachineState(int NewState);
     int currentState=0;
     int TransitionState=0;
-    unsigned long TransitionStartMicros=0;
+    uint32_t TransitionStartMicros=0;
     int TelemetryMode=1;
-    unsigned long telemMillis=0;
-    unsigned long telemMillisRefreshRate=1000;
+    uint32_t telemMillis=0;
+    uint32_t telemMillisRefreshRate=1000;
 };
